Update the setup to remove the step to configure docker
Step 4 in the installation instructions ask the user to create a docker config file and add the contents manually. However, the docker config file docker-compose.yml is already included in the stable build, so this step is redundant. In fact, this step is wrong because if you follow this step and update the contents of the config file, the next step "docker-compose build: will fail. So this erroneous step needs to be removed or replaced with a correct step. This PR modifies the instruction to verify that the docker-compose config file exists in the folder.